pyproject.toml
[tool.poetry]
name = "storageinator"
version = "0.1.0"
description = "File storage service with S3 backend"
authors = ["artykovv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
motor = "^3.7.1"
boto3 = "^1.42.21"
python-multipart = "^0.0.21"
python-dotenv = "^1.2.1"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
aiofiles = "^25.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
email-validator = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
